+++ release/src/components/button/index.d.ts
@@ -1,0 +1,2 @@
+export * from './button.module';
+export * from './file-button.component';