+++ release/src/components/loading/loading.component.d.ts
@@ -1,0 +1,4 @@
+export declare class LoadingComponent {
+    visible: boolean;
+    progress: number;
+}