+++ release/src/components/notification/notification.component.d.ts
@@ -1,0 +1,22 @@
+import { EventEmitter } from '@angular/core';
+import { NotificationService } from './notification.service';
+import { NotificationStyleType } from './notification-style.type';
+export declare class NotificationComponent {
+    private notificationService;
+    cssClass: string;
+    title: string;
+    body: string;
+    template: any;
+    pauseOnHover: boolean;
+    styleType: NotificationStyleType;
+    showClose: boolean;
+    timestamp: any;
+    close: EventEmitter<{}>;
+    pause: EventEmitter<{}>;
+    resume: EventEmitter<{}>;
+    timeout: any;
+    readonly cssClasses: string;
+    constructor(notificationService: NotificationService);
+    onMouseEnter(): void;
+    onMouseLeave(): void;
+}