+++ release/src/components/tabs/index.d.ts
@@ -1,0 +1,3 @@
+export * from './tabs.module';
+export * from './tabs.component';
+export * from './tab.component';